clean-up/renaming to match style guide
