+++ nba-front-end/src/components/PlayerPage/PlayerPage.tsx
@@ -2,9 +2,9 @@ import * as React from 'react';
 import Container from '@mui/material/Container';
 import Grid from '@mui/material/Grid';
 import Paper from '@mui/material/Paper';
-import FilledPlayerTable from './playerTableLoader';
+import PlayerTableLoader from './PlayerTableLoader';
 
-function PlayerDataGridPage() {
+function PlayerPage() {
   return (
     //   these lines set up the format of the page
     <Container maxWidth="lg" sx={{ mt: 4, mb: 4, "min-height":440 }}>
@@ -12,11 +12,11 @@ function PlayerDataGridPage() {
             {/* Chart */}
             <Grid item xs={12} md={12} lg={12}>
                 {/* this is the table of players, see playerTableLoader.tsx */}
-                <FilledPlayerTable></FilledPlayerTable>
+                <PlayerTableLoader></PlayerTableLoader>
             </Grid>
         </Grid>
     </Container>
   );
 }
 
-export default PlayerDataGridPage;
+export default PlayerPage;
