+++ nba-front-end/src/components/PlayerPage/PlayerTableLoader.tsx
@@ -1,18 +1,18 @@
 import React, { useEffect, useState } from 'react';
 // import { Observable, of, Subject } from 'rxjs';
 // import { catchError, reduce, take } from 'rxjs/operators';
-import DataGridPlayers from './addPlayerDataGrid';
+import PlayerDataGrid from './PlayerDataGrid';
+import ApiComponentLoader from '../ApiComponentLoader';
 import api from '../../services/api';
 import { Player } from '../../models/IPlayer';
-
+import { throwError, of } from 'rxjs';
+import { catchError } from 'rxjs/operators';
 
 interface PlayerProps{
     playerList: Player[];
 }
 
-const FilledAddPlayerTable: React.FC<any> = (props) => {
-
-  const teamID = props.teamID;
+function PlayerTableLoader() {
   // this sets up function (from componentLoading.tsx) which either returns 
   // DataGridPlayers (playerDataGrid.tsx) 
   // or
@@ -20,7 +20,7 @@ const FilledAddPlayerTable: React.FC<any> = (props) => {
   // we call this as <TableLoading isLoading={appState.loading} playerList={appState.playerList} />
   // isLoading={appState.loading} says that the boolean for whether data is available is 'appState.loading' variable
   // playerList={appState.playerList} says that (if the data is available) send the appState.playerList to the DataGridPlayers Component
-  const TableLoading = (DataGridPlayers);
+  const TableLoading = (PlayerDataGrid);
   const [appState, setAppState] = useState<PlayerProps>({
     playerList: [],
   });
@@ -28,7 +28,6 @@ const FilledAddPlayerTable: React.FC<any> = (props) => {
   const [errorMessage, setErrorMessage] = useState("");
     // defines a state for when the api is fetching data for players
   const [isLoading, setLoading] = useState(false);
-
 
   // this is the call to the API to get the player data
   useEffect(() => {
@@ -53,12 +52,10 @@ const FilledAddPlayerTable: React.FC<any> = (props) => {
   {/* if the error message is not empty or does not equal "", then the error message will appear*/}
         {errorMessage!==""&&<h1 style={{color: 'red'}}>Oops! An Error Occured Please Try Again.</h1>}
   {/* if  isLoading is true, loading text will apear, if api is able to fetch player data and isLoading is false, then show filled player table*/}
-        {isLoading ? (<h1>Hold on, fetching data may take some time :)</h1>) : (<TableLoading playerList={appState.playerList} teamID={teamID} 
-        tableIsUpdated={props.tableIsUpdated}
-        />)}
+        {isLoading ? (<h1>Hold on, fetching data may take some time :)</h1>) : (<TableLoading playerList={appState.playerList} />)}
       </div>
     </React.Fragment>
 
   );
 }
-export default FilledAddPlayerTable;
+export default PlayerTableLoader;
