+++ nba-front-end/src/components/TeamPage/AddPlayerTableLoader.tsx
@@ -1,26 +1,26 @@
 import React, { useEffect, useState } from 'react';
 // import { Observable, of, Subject } from 'rxjs';
 // import { catchError, reduce, take } from 'rxjs/operators';
-import DataGridPlayers from './playerDataGrid';
-import WithTableLoading from '../componentLoading';
+import AddPlayerTable from './AddPlayerTable';
 import api from '../../services/api';
 import { Player } from '../../models/IPlayer';
-import { throwError, of } from 'rxjs';
-import { catchError } from 'rxjs/operators';
+
 
 interface PlayerProps{
     playerList: Player[];
 }
 
-function FilledPlayerTable() {
+const AddPlayerTableLoader: React.FC<any> = (props) => {
+
+  const teamID = props.teamID;
   // this sets up function (from componentLoading.tsx) which either returns 
-  // DataGridPlayers (playerDataGrid.tsx) 
+  // AddPlayerTable (playerDataGrid.tsx) 
   // or
   // a message 'Hold on, fetching data may take some time :)'
   // we call this as <TableLoading isLoading={appState.loading} playerList={appState.playerList} />
   // isLoading={appState.loading} says that the boolean for whether data is available is 'appState.loading' variable
-  // playerList={appState.playerList} says that (if the data is available) send the appState.playerList to the DataGridPlayers Component
-  const TableLoading = (DataGridPlayers);
+  // playerList={appState.playerList} says that (if the data is available) send the appState.playerList to the AddPlayerTable Component
+  const TableLoading = (AddPlayerTable);
   const [appState, setAppState] = useState<PlayerProps>({
     playerList: [],
   });
@@ -28,6 +28,7 @@ function FilledPlayerTable() {
   const [errorMessage, setErrorMessage] = useState("");
     // defines a state for when the api is fetching data for players
   const [isLoading, setLoading] = useState(false);
+
 
   // this is the call to the API to get the player data
   useEffect(() => {
@@ -52,10 +53,12 @@ function FilledPlayerTable() {
   {/* if the error message is not empty or does not equal "", then the error message will appear*/}
         {errorMessage!==""&&<h1 style={{color: 'red'}}>Oops! An Error Occured Please Try Again.</h1>}
   {/* if  isLoading is true, loading text will apear, if api is able to fetch player data and isLoading is false, then show filled player table*/}
-        {isLoading ? (<h1>Hold on, fetching data may take some time :)</h1>) : (<TableLoading playerList={appState.playerList} />)}
+        {isLoading ? (<h1>Hold on, fetching data may take some time :)</h1>) : (<TableLoading playerList={appState.playerList} teamID={teamID} 
+        tableIsUpdated={props.tableIsUpdated}
+        />)}
       </div>
     </React.Fragment>
 
   );
 }
-export default FilledPlayerTable;
+export default AddPlayerTableLoader;
