+++ nba-front-end/src/components/TeamPage/TeamPage.tsx
@@ -1,0 +1,14 @@
+import * as React from "react";
+import Container from "@mui/material/Container";
+import TeamPageContentLoader from "./TeamPageContentLoader";
+
+function TeamPage() {
+  return (
+    <Container maxWidth="lg" sx={{ mt:4, mb: 4, "minHeight": '600px' }}>
+      {/* Chart */}
+      <TeamPageContentLoader/>
+    </Container>
+  );
+}
+
+export default TeamPage;
