+++ nba-front-end/src/routes/routes.tsx
@@ -2,11 +2,10 @@ import React from "react";
 import { Routes, Route } from "react-router-dom";
 import BasePage from "../components/Base";
 import DefaultPage from "../components/Default";
-import LoginPage from "../components/loginPage/LoginPage";
+import LoginPage from "../components/LoginPage/LoginPage";
 import Dashboard from "../components/Menu";
-import PlayerDataGridPage from '../components/playerDataGrid/playerDataGridPage';
-import TeamDataGridPage from '../components/teamDataGrid/teamDataGridPage'
-import TeamsPage from "../components/teamDataGrid/teamDataGridPage";
+import PlayerPage from '../components/PlayerPage/PlayerPage';
+import TeamPage from "../components/TeamPage/TeamPage";
 
 function RouteConfig() {
   return (
@@ -15,8 +14,8 @@ function RouteConfig() {
             <Route path="/" element={<DefaultPage />} />
             <Route path="login" element={<LoginPage />} />
             <Route path="dashboard" element={<Dashboard />}>
-                <Route path="players" element={<PlayerDataGridPage />} />
-                <Route path="teams" element={<TeamDataGridPage />} />
+                <Route path="players" element={<PlayerPage />} />
+                <Route path="teams" element={<TeamPage />} />
             </Route>
             {/* Using path="*"" means "match anything", so this route
                 acts like a catch-all for URLs that we don't have explicit
