+++ nba-front-end/src/services/axios_config.ts
@@ -1,7 +1,7 @@
 import { AxiosRequestConfig } from 'axios';
 
 export const axiosRequestConfiguration: AxiosRequestConfig = {
-    baseURL: `${process.env.REACT_APP_AzureAPIDotNet}`,
+    baseURL: `${process.env.REACT_APP_LocalAPIDotNet}`,
     responseType: 'json',
     headers: {
         'Content-Type': 'application/json',
